pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
discord = "^2.2.3"
bs4 = "^0.0.1"
Pillow = "^9.5.0"
py-fumen-util = "0.3.0"
praw = "^7.7.0"
lzstring = "^1.0.4"
matplotlib = "^3.7.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
"discord.py" = "^2.3.2"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"